Remove IsRebase from CompressImagesMessage. Remove Update from OpenPrMessage. Move ParseCommitMessage to CommitMessage.cs
