+++ Common/Messages/OpenPrMessage.cs
@@ -7,7 +7,5 @@
         public string RepoName { get; set; }
 
         public string CloneUrl { get; set; }
-
-        public bool Update { get; set; }
     }
 }
