+++ CompressImagesFunction/CompressImagesFunction.cs
@@ -102,12 +102,6 @@ namespace CompressImagesFunction
                 RepoOwner = compressImagesMessage.Owner,
             });
 
-            if (branchExists && !compressImagesMessage.IsRebase)
-            {
-                logger.LogInformation("CompressImagesFunction: skipping repo {Owner}/{RepoName} as branch exists", compressImagesMessage.Owner, compressImagesMessage.RepoName);
-                return;
-            }
-
             var compressImagesParameters = new CompressimagesParameters
             {
                 CloneUrl = compressImagesMessage.CloneUrl,
@@ -115,7 +109,7 @@ namespace CompressImagesFunction
                 Password = installationToken.Token,
                 RepoName = compressImagesMessage.RepoName,
                 RepoOwner = compressImagesMessage.Owner,
-                IsRebase = compressImagesMessage.IsRebase,
+                IsRebase = branchExists,
                 PgpPrivateKey = KnownEnvironmentVariables.PGP_PRIVATE_KEY,
                 PgPPassword = KnownEnvironmentVariables.PGP_PASSWORD,
                 CompressImagesMessage = compressImagesMessage,
@@ -130,14 +124,12 @@ namespace CompressImagesFunction
             }
             else if (didCompress)
             {
-                var update = compressImagesMessage.IsRebase;
                 logger.LogInformation("CompressImagesFunction: Successfully compressed images for {Owner}/{RepoName}", compressImagesMessage.Owner, compressImagesMessage.RepoName);
                 openPrMessages.Add(new OpenPrMessage
                 {
                     InstallationId = compressImagesMessage.InstallationId,
                     RepoName = compressImagesMessage.RepoName,
                     CloneUrl = compressImagesMessage.CloneUrl,
-                    Update = compressImagesMessage.IsRebase,
                 });
             }
