+++ OpenPrFunction/IPullRequest.cs
@@ -6,6 +6,6 @@ namespace OpenPrFunction
 {
     public interface IPullRequest
     {
-        Task<Pr> OpenAsync(GitHubClientParameters parameters, bool update, Settings settings = null);
+        Task<Pr> OpenAsync(GitHubClientParameters parameters, Settings settings = null);
     }
 }
