+++ OpenPrFunction/PullRequest.cs
@@ -9,7 +9,7 @@ namespace OpenPrFunction
 {
     public class PullRequest : IPullRequest
     {
-        public async Task<Pr> OpenAsync(GitHubClientParameters parameters, bool update, Settings settings = null)
+        public async Task<Pr> OpenAsync(GitHubClientParameters parameters, Settings settings = null)
         {
             var inMemoryCredentialStore = new InMemoryCredentialStore(new Credentials(KnownGitHubs.Username, parameters.Password));
             var githubClient = new GitHubClient(new ProductHeaderValue("ImgBot"), inMemoryCredentialStore);
@@ -30,39 +30,13 @@ namespace OpenPrFunction
             }
 
             var stats = Stats.ParseStats(commit.Commit.Message);
+        
+            var pr = new NewPullRequest(KnownGitHubs.CommitMessageTitle, KnownGitHubs.BranchName, baseBranch)
+            {
+                Body = PullRequestBody.Generate(stats),
+            };
 
-            Octokit.PullRequest result;
-            if(update) {
-                //get PR number
-                var allPrs = await githubClient.PullRequest.GetAllForRepository(parameters.RepoOwner, parameters.RepoName);
-                
-                var prNumber = 0;
-                foreach(var pr in allPrs) {
-                    if (pr.State == ItemState.Open && pr.Head.Sha == branch.Commit.Sha)
-                    {
-                        prNumber = (int)pr.Number;
-                    }
-                }
-
-                if (prNumber == 0) {
-                    throw new Exception("Couldn't update PR. PR not found");
-                }
-
-                var pru = new PullRequestUpdate()
-                {
-                    Body = PullRequestBody.Generate(stats),
-                };
-
-                result = await githubClient.PullRequest.Update(parameters.RepoOwner, parameters.RepoName, prNumber, pru);
-            }
-            else  {
-                var pr = new NewPullRequest(KnownGitHubs.CommitMessageTitle, KnownGitHubs.BranchName, baseBranch)
-                {
-                    Body = PullRequestBody.Generate(stats),
-                };
-
-                result = await githubClient.PullRequest.Create(parameters.RepoOwner, parameters.RepoName, pr);
-            }
+            Octokit.PullRequest result = await githubClient.PullRequest.Create(parameters.RepoOwner, parameters.RepoName, pr);
 
             if (stats == null)
             {
