+++ Test/CommitMessageTests.cs
@@ -291,7 +291,7 @@ namespace Test
                 },
             };
 
-            var parsed = CompressionResult.ParseCommitMessage(commitMessage);
+            var parsed = CommitMessage.ParseCommitMessage(commitMessage);
             Assert.AreEqual(parsed.Length, expected.Length);
 
             for (int i = 0; i < parsed.Length; ++i)
