+++ WebHook/WebHookFunction.cs
@@ -111,7 +111,6 @@ namespace WebHook
                     InstallationId = hook.installation.id,
                     RepoName = hook.repository.name,
                     CloneUrl = $"https://github.com/{hook.repository.full_name}",
-                    Update = false,
                 })));
 
                 logger.LogInformation("ProcessPush: Added OpenPrMessage for {Owner}/{RepoName}", hook.repository.owner.login, hook.repository.name);
